pyproject.toml
[project]
name = "jsonpatch-trigger"
version = "0.1.0"
description = "Extension for JSONPatch (RFC6902) that introduces operation preconditions, change tracking and automated listeners that can introduce new operations based on the previous change"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-jsonpath==2.0.2",
    "pydantic==2.12.5"
]
[project.optional-dependencies]
test = [
    "coverage==7.13.4",
    "pytest==9.0.2"
]